dashboard_back/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
idna==3.7
iniconfig==2.0.0
openpyxl==3.1.2
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.1
requests==2.32.2
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.0
urllib3==2.2.1
